+++ packages/ethereum-contracts/contracts/superfluid/SuperTokenFactory.sol
@@ -212,7 +212,7 @@ abstract contract SuperTokenFactoryBase is
         Upgradability upgradability,
         string calldata name,
         string calldata symbol,
-        address adminOverride
+        address admin
     ) public override returns (ISuperToken superToken) {
         if (address(underlyingToken) == address(0)) {
             revert SUPER_TOKEN_FACTORY_ZERO_ADDRESS();
@@ -232,12 +232,12 @@ abstract contract SuperTokenFactoryBase is
         }
 
         // initialize the token
-        superToken.initializeWithAdminOverride(
+        superToken.initializeWithAdmin(
             underlyingToken,
             underlyingDecimals,
             name,
             symbol,
-            adminOverride
+            admin
         );
 
         emit SuperTokenCreated(superToken);
@@ -270,7 +270,7 @@ abstract contract SuperTokenFactoryBase is
         Upgradability upgradability,
         string calldata name,
         string calldata symbol,
-        address adminOverride
+        address admin
     )
         external override
         returns (ISuperToken superToken)
@@ -281,7 +281,7 @@ abstract contract SuperTokenFactoryBase is
             upgradability,
             name,
             symbol,
-            adminOverride
+            admin
         );
     }