+++ packages/sdk-core/src/SuperToken.ts
@@ -1,6 +1,8 @@
 import { BytesLike, ethers, Overrides } from "ethers";
 
 import ConstantFlowAgreementV1 from "./ConstantFlowAgreementV1";
+import ConstantInflowNFT from "./ConstantInflowNFT";
+import ConstantOutflowNFT from "./ConstantOutflowNFT";
 import ERC20Token from "./ERC20Token";
 import Governance from "./Governance";
 import InstantDistributionAgreementV1 from "./InstantDistributionAgreementV1";
@@ -52,7 +54,13 @@ import {
     getSanitizedTimestamp,
     getStringCurrentTimeInSeconds,
     normalizeAddress,
+    tryGet,
 } from "./utils";
+
+export interface NFTAddresses {
+    readonly constantInflowNFTProxy: string;
+    readonly constantOutflowNFTProxy: string;
+}
 
 export interface ITokenSettings {
     readonly address: string;
@@ -81,6 +89,10 @@ export default abstract class SuperToken extends ERC20Token {
     readonly idaV1: InstantDistributionAgreementV1;
     readonly governance: Governance;
     readonly underlyingToken?: ERC20Token;
+    readonly constantOutflowNFTProxy?: ConstantOutflowNFT;
+    readonly constantInflowNFTProxy?: ConstantInflowNFT;
+    readonly constantOutflowNFTLogic?: string;
+    readonly constantInflowNFTLogic?: string;
     override readonly contract: ISuperToken;
 
     protected constructor(options: ITokenOptions, settings: ITokenSettings) {
@@ -124,9 +136,13 @@ export default abstract class SuperToken extends ERC20Token {
                 options.address,
                 options.provider
             );
-            const underlyingTokenAddress = await superToken
+            const getUnderlyingTokenPromise = superToken
                 .connect(options.provider)
                 .getUnderlyingToken();
+            const underlyingTokenAddress = await tryGet(
+                getUnderlyingTokenPromise,
+                ethers.constants.AddressZero
+            );
             const settings: ITokenSettings = {
                 address: options.address,
                 config: options.config,
@@ -146,21 +162,73 @@ export default abstract class SuperToken extends ERC20Token {
             const nativeTokenSymbol = resolverData.nativeTokenSymbol || "ETH";
             const nativeSuperTokenSymbol = nativeTokenSymbol + "x";
 
+            // @note This is tech debt and should be reverted once GoodDollar upgrades their token contract
+            // @note We are using tryGet here just to handle GoodDollar not having
+            // CONSTANT_OUTFLOW_NFT in its SuperToken implementation.
+            let constantOutflowNFTProxy = await tryGet(
+                superToken.CONSTANT_OUTFLOW_NFT(),
+                ethers.constants.AddressZero
+            );
+            let constantInflowNFTProxy = await tryGet(
+                superToken.CONSTANT_INFLOW_NFT(),
+                ethers.constants.AddressZero
+            );
+
+            // @note We only want to run this bit of code for GoodDollar SuperTokens
+            // (dev and mainnet)
+            const GOOD_DOLLAR_SYMBOL = "G$";
+            if (tokenSymbol === GOOD_DOLLAR_SYMBOL) {
+                // @note we need to create a new interface for the old GoodDollar SuperToken
+                // which contains the functions for constantInflowNFT and constantOutflowNFT
+                const oldSuperTokenInterface = new ethers.utils.Interface([
+                    "function constantInflowNFT() view returns (address)",
+                    "function constantOutflowNFT() view returns (address)",
+                ]);
+                const goodDollarSpecificToken = new ethers.Contract(
+                    superToken.address,
+                    oldSuperTokenInterface
+                );
+
+                // @note we attempt to get the constantInflowNFT and constantOutflowNFT
+                if (constantOutflowNFTProxy === ethers.constants.AddressZero) {
+                    constantOutflowNFTProxy = await tryGet(
+                        goodDollarSpecificToken.constantOutflowNFT(),
+                        ethers.constants.AddressZero
+                    );
+                }
+                if (constantInflowNFTProxy === ethers.constants.AddressZero) {
+                    constantInflowNFTProxy = await tryGet(
+                        goodDollarSpecificToken.constantInflowNFT(),
+                        ethers.constants.AddressZero
+                    );
+                }
+            }
+
+            const nftAddresses: NFTAddresses = {
+                constantOutflowNFTProxy,
+                constantInflowNFTProxy,
+            };
+
             if (nativeSuperTokenSymbol === tokenSymbol) {
                 return new NativeAssetSuperToken(
                     options,
                     settings,
-                    nativeTokenSymbol
+                    nativeTokenSymbol,
+                    nftAddresses
                 );
             }
 
             if (underlyingTokenAddress !== ethers.constants.AddressZero) {
-                return new WrapperSuperToken(options, {
-                    ...settings,
-                    underlyingTokenAddress,
-                });
+                return new WrapperSuperToken(
+                    options,
+                    {
+                        ...settings,
+                        underlyingTokenAddress,
+                    },
+                    nftAddresses
+                );
             }
-            return new PureSuperToken(options, settings);
+            return new PureSuperToken(options, settings, nftAddresses);
         } catch (err) {
             throw new SFError({
                 type: "SUPERTOKEN_INITIALIZATION",
@@ -746,13 +814,22 @@ export default abstract class SuperToken extends ERC20Token {
  */
 export class WrapperSuperToken extends SuperToken {
     override readonly underlyingToken: ERC20Token;
+    override readonly constantOutflowNFTProxy: ConstantOutflowNFT;
+    override readonly constantInflowNFTProxy: ConstantInflowNFT;
 
     constructor(
         options: ITokenOptions,
-        settings: ITokenSettings & { underlyingTokenAddress: string }
+        settings: ITokenSettings & { underlyingTokenAddress: string },
+        nftAddresses: NFTAddresses
     ) {
         super(options, settings);
         this.underlyingToken = new ERC20Token(settings.underlyingTokenAddress);
+        this.constantInflowNFTProxy = new ConstantInflowNFT(
+            nftAddresses.constantInflowNFTProxy
+        );
+        this.constantOutflowNFTProxy = new ConstantOutflowNFT(
+            nftAddresses.constantOutflowNFTProxy
+        );
     }
 
     /** ### WrapperSuperToken Contract Write Functions ### */
@@ -854,8 +931,21 @@ export class WrapperSuperToken extends SuperToken {
  * PureSuperToken doesn't have any underlying ERC20 token.
  */
 export class PureSuperToken extends SuperToken {
-    constructor(options: ITokenOptions, settings: ITokenSettings) {
+    override readonly constantOutflowNFTProxy: ConstantOutflowNFT;
+    override readonly constantInflowNFTProxy: ConstantInflowNFT;
+
+    constructor(
+        options: ITokenOptions,
+        settings: ITokenSettings,
+        nftAddresses: NFTAddresses
+    ) {
         super(options, settings);
+        this.constantInflowNFTProxy = new ConstantInflowNFT(
+            nftAddresses.constantInflowNFTProxy
+        );
+        this.constantOutflowNFTProxy = new ConstantOutflowNFT(
+            nftAddresses.constantOutflowNFTProxy
+        );
     }
 }
 
@@ -864,13 +954,23 @@ export class PureSuperToken extends SuperToken {
  */
 export class NativeAssetSuperToken extends SuperToken {
     readonly nativeTokenSymbol: string;
+    override readonly constantOutflowNFTProxy: ConstantOutflowNFT;
+    override readonly constantInflowNFTProxy: ConstantInflowNFT;
+
     constructor(
         options: ITokenOptions,
         settings: ITokenSettings,
-        nativeTokenSymbol: string
+        nativeTokenSymbol: string,
+        nftAddresses: NFTAddresses
     ) {
         super(options, settings);
         this.nativeTokenSymbol = nativeTokenSymbol;
+        this.constantInflowNFTProxy = new ConstantInflowNFT(
+            nftAddresses.constantInflowNFTProxy
+        );
+        this.constantOutflowNFTProxy = new ConstantOutflowNFT(
+            nftAddresses.constantOutflowNFTProxy
+        );
     }
 
     get nativeAssetContract() {